Creation of the endpoint /login/
when the router's urls are include into the url patterns :
urlpatterns = [
	path('', include(router.urls)),
    path('admin/', admin.site.urls),
    path('login/', TokenObtainPairView.as_view(), name='obtain_tokens'),
    path('login/refresh/', TokenRefreshView.as_view(), name='refresh_token'),
]
there lead to a conflit on the endpoint /login/. we receive the message that this endpoint
are not allowing access through post methods. But with /api/login/ the endpoint work as expected.

by adding the router's url after urlpatterns :
urlpatterns += router.urls
the endpoint /login/ is completly fixed.